/ rc: fix the version check to actually work
it only seemed to work because it had the right behavior on 2/3 systems
tested.  still don't know the full issue but there's some escaping going
on and apparently parentheses are problematic, this might only be using
system(3) which is going to be /bin/sh and therefore doesn't have things
like arithmetic expansion.

this is only explanation, but for the moment just want it to work, so
got it to that point and moving on, maybe figure it out more later.