- rc: make it use the new v2 config also for v2.9a
it confuses the pattern.  could fix that, but we can't use the
greater-than then (or have to do something special with letters
afterwards, which opens a can of worms)

so just do this quick by comparing directly to that tag.  there's no
other 2.9x and 3.x is already out so unlikely to be more 2.9x