- rc: lower scrollback to speed new-window and refresh-client -S
see tmux issue 3352, this is the best-effort workaround for now until we
can try to implement the mentioned patch idea to avoid duplication of
page tables in all forked processes.  this is consequence of single
server process design of tmux, large page table mappings take a while to
duplicate.  maybe they are being duplicated imperatively rather than
just memcopies because they cannot be copied directly in that context?
not sure how else to explain the slow speed even a few hundred MB of
pages should not take so long to copy, but it does.

so, lowering from ~2B to ~1M scrollback lines limit after doing some
performance testing.  this is 1M at 80chars, of course not all lines
will be that long and some will be longer, hopefully tmux will not have
to keep dummy data for chars which aren't drawn, don't know its
internals enough to say.

in any case this should help the persistent performance issues present
in task a303b4b8