/ rc: make version test work even with multiple tmux versions
this fixes but where tmux invoked with a different version (for testing,
during a build, etc) was using the wrong binary on the system for
version test.

we were just running "tmux" itself alone, which can end up resolving to
something different than the invocation executable (and in general,
should not go through any wrapping probably, since it's already itself
in a wrapper and doing something very simple just talking to the server
process and exiting), and thus load the wrong configuration file

instead we parse tmux pid from $TMUX and use it to look up the actual
executable used via /proc filesystem.  obviously this hack will only
work on linux, this should be ok, unless we end up using this on mac
some day, which we don't want to do and will fix when the need arises