+ v2rc: left: run git only once for statchars and handle untracked
this avoids another expensive fork(), see tmux issue 3352 for why we
want to avoid as many subshells as possible

fortunately tmux re-interprets the string it gets for expansion
characters, so we are able to print the color prefix ourselves

note that this version also handles '?' chars (ie untracked files),
which were not handled before.  since they were already inline this was
easy to just implement in the first pass, have been wanting this for
some time.  kept it with color #[default] which should be somewhat muted
compared to the others