+ v2rc: make the initial session always named '/'
there are better ways to do this in later tmux, but this way seems to
work even on older v2.  when parsing the config file, session does not
exist yet, so we have to hook on any session being created (which will
include the first one).  when run in first session init, we destroy the
hook so it only will ever run that once.  this could be race-prone, but
there's only one session at start and others are only created manually
by keybind, so it probably won't come up TODO