+ v2rc: bind: vi keys for win/pane/session up/down/l/r
capital-vi for pane travel, lower-vi does sessions on one dimension and
windows on the other.  then last window is by prefix-space and
last-session is prefix-control-space, this keeps it like ratpoison which
is switching with meta-space and then control-space switches groups
(vscreens).  so all the dimensions change to the last one with the key
that accesses that dimension, plus a space