/ v2rc: bind: assign B as choose-buffer, compensates winmv '='
we stole '=' (by default, mapped to "choose-buffer -Z") to use it for
window movement functions in commit d5c7f89 (to correspond to the +/-
caps really, since we are doing +/- movements for windows in that
commit).  so here we give a key back for it, not that we ever used it,
but actually it's pretty interesting when trying it out! should think
more about uses...