- v2rc: rewrite some idioms to work on older 2.x tmux
this config worked fine on next-3.4, but not eg 3.1a (u22) or the older
manual builds of 2.9-next we used.  some slight modifications were
needed to quoting and command queue creation; we also set the user vars
@latest, @major, @minor, @version to be used later down the road, these
will let us use conditionals for more advanced concepts later.
unfortunately the full conditional syntax is not supported in early 2.x
but we can maybe have a follow-on v3rc and source that? TODO