- v2rc: make all features work on 2.6, 3.0a, 3.1c, 3.2a, next-3.4
this required much testing and rethinking implementations of stuff to be
compatible (syntax was much easier to "predict" with next-3.4),
especially expansion stuff is very tricky, but this seems to be
compatible now with the whole set of features that we use

(except 1.6, which loads v1rc out of rc (tmux.conf))

in some cases we needed different implementations, eg change in behavior
whether pane swap auto-last-panes or not between versions, different
flags, which we used conditional aliases, things like that.  differences
are fairly readable and easy to reproduce / amend later

in particular this make window move, window first and end, keys that
fully create the /, -, * sessions, including differences in pane
percentage setting.  then also we have good framework for changes later
and continued maintenance with @latest, @major, @minor, @version
variables