- alias: more indirect command aliases to simplify binds
condense the command chains with these indirections, they are useful to
make it more compact

also re-layout existing related alias groups offset by 10 in this
commit, we had used 5 before but the gap is not enough.  we can see
we'll be doing this often.  this is unfortunate that we have to maintain
these numbers because later tmux (3.4 works) seems to be fine with no
specified index.  maintaining them is required if we want to maintain
compatibility back to 2.6 though.  sad but true... no way to optimize
this because it would just be duplicated even if we did it the right way
in parallel