+ v2rc: bind: goto /, -, *, but no recreate wins if exist
this means we can use them as fixed-session switch keys, and if the
session does not exist already it will be created.  depends on (1) later
tmux (not sure what version, but we are using @latest as version check
and it's in 3.2 docs)

on versions that don't support the necessary expansion, we have to just
do it anyways, because it could be the first time.  so user will just
have to close those windows if he didn't mean it.  anyways, we will
probably never encounter this as we'll always be doing non-trivial stuff
on self-built tmux versions anyways