- v2rc: bind: don't do full pane init macros if not on latest tmux
we have some special sessions that we always set up the same on start,
and rely on the first switch to the session (using special keybind
particular to that session) to create the panes if they don't exist.

previously, if testing for window existence by name was unsupported in
the tmux version (ie, if @latest == 0), we just would blindly create all
the pre-made session windows over and over again every time the key was
pressed (thereby creating duplicate windows).  the behavior of course is
supposed to be that the session windows are created only the very first
keypress.

since there's no way to test for existence of a window by name on early
tmux than 3.2, we had previously opted to keep the binds
unconditionally, but this really doesn't make sense because (1) we will
always have latest tmux where we make these special sessions, ie only on
our local workstation, and (2) these sessions wouldn't have any need to
exist on remote systems, at least with all the diferent windows and
layout that we would use on the workstation.  we still keep the ability
to make and change to the sessions but we won't touch the windows
inside.