/ v2rc: status: we use gawk-isms and some systems have mawk as awk
mawk doesn't support some things like rshift() so it will not work

actually if you install gawk it becomes the alternative for mawk, but
just using 'awk' caused me to be pretty confused on a system where gawk
wasn't yet installed.  debatable whether this change will cause easier
debugging because it will be an empty value in either case and hidden
from the user but at least it will be apparent on trace that gawk
doesn't exist rather than having to try to duplicate the command.  then
once installed it will magically start working...

gawk is already in setup repo for a standard node so this only came up
because we hadn't run full cfmgmt yet on a certain system where we are
testing tmux