+ tmux: add wrapper that sets THP malloc tunable, faster pane starts
wrapper for starting tmux that sets glibc environment variable for
malloc to use huge pages.  see explanation in tmux rcfile.

this script needs to be installed with installx, and not be a symlink.
it replaces src/.bash@dbe32d99 which put this in a shell function, but
it's better to have it in a script wrapper.