/ v2rc: bind: reduce version complexity of swap window binds
tested this and the behavior of pre-3.0 seems to be the same as post-3.0
except it swaps meaning of '-d', otherwise they work the same.  not sure
why we had a special case for 3.0 itself

we can successfully use '>= 3' with older versions, because even though
the >= expansion is unsupported until 3.0, it seems to always evalute
to 0 -- without an error -- earlier than that.  so the effect is the
same as if it were supported