- v2rc: bind: do version test for @latest early to short-circuit
on non-@latest, this will result in one less test and avoid even having
to try the N/w evaluation, even though it works correctly, this makes it
more clear that the version is needed for the expansion.  this relies on
tmux's '&&' doing short-circuit eval, which tbh haven't checked, but
probably it does.  if not, no harm done since the expansion does work
correctly even when not supported (ie it evaluates to false).  this
behavior was a wise choice by tmux devs to make it forwards-compatible

note, we also stopped explicitly checking for the value '1' because that
is not necessary (if expands to '0' it evaluates as false, tested this)