- v2rc: version: use simpler @{vMN} "at least" semantics for variants
we now set a @vX where X is the <major><minor> version that supports
specific features, and in the case that it's derived from #{version}, is
done after stripping non-digits alphabetics.  in this way the minimum
version for a particular stanza can be asserted with simple test for "at
least @v26" meaning if @v26 is set to 1, the version is at least that
much.  however only certain variables we care about are actually
defined, for example there is a v2.8 but nothing we care about uses
anything made in v2.8, so we only have @v26, @v27, then @v30.  should be
clear from looking at the version definition stanza

this gets rid of @latest, which is rather non-deterministic

with these vars, we were also able to simplify some stanza that differed
by only a few characters or less by using "#{?#{@vX},<yesval>,<noval>}"
construct

overall this should make for simpler rcfiles that keep version checks in
one section and don't put any logic elsewhere

the v2rc file is actually only sourced for 2.6+ but nonetheless we add
some dummy @vX variables to placehold the information about important
features that were added in particular versions, since we have now
researched this by looking directly at commits and don't want to repeat
the process since it took a long time

also note that for some key binds, we had to add alias indirection,
because 'bind' does not expand operands as formats and does not have a
'-F' to do it.  so we have to expand it in the alias definition and then
call the alias.  it would be nice if someone added a '-F' to 'bind'
(TODO task 1b9f307a)