+ rc: version: replace awk-based v2.6 test with pure /bin/sh impl
this only has one fork, the call to tmux -V.  tested with /bin/sh as
dash.  this removes the need for gawk or bq.  it also looks more
straightforward.  we did get rid of 'master' but after building lots of
versions 1.6-3.4, did not see this.  might happen if we use our own
patch though, as we did at one time for testing bugfix 1627.  this is
expected to be a rare event we can figure out if it comes up again.