- v2rc: status: run lockless when we call out to "git status"
git status normally updates the cache as an optimization for the next
run.  but for things running in the background (like the tmux
refresh-client -S that we do in bash prompt_command()), this just causes
problems with the foreground terminal running git commit, etc.  even
though it doesn't happen very often, it has happened enough to be
annoying.

there is a cli flag for this (--no-optional-locks) but by using
environment, we do not have to worry about what version of git is
installed or what version has the feature, it just won't be enabled if
unavailable in the system's git version