- tmux: env: glibc: malloc: set more tunables to try to force mmap
set these for lowest possible because hugetlb=1 only comes into play
when malloc uses mmap.  otherwise in heap case hugetlb will not be used.

tried jemalloc, but it does not work with static linking.  because it's
in the linker that the calls are fixed up apparently.  at least, could
not personally get it to work.  could also be something as simple as the
link line order, jemalloc may not have been first, before the implicit
libc, but we'll have to leave that investigation for later as there is
no time currently

these are only logically the lowest possible rather than confirmed by
looking at the allowed range in the source code TODO