- tmux: client: use 'exec' to avoid an lingering extra bash on attach
this keeps the process table clean.  it shouldn't matter where we came
from, that we were started by bash, and there is no reason for bash to
hang around waiting for tmux to finish, only to pass along the exit
code, we can just use exec, as unix intends.

this was already done for the server, just not the client.