- tmux: make argv[0] just 'tmux' rather than 'exe' or the path
for the client, obviously we don't want to be called 'exe' in the
process table.  we don't want to readlink though because that incurs
extra expense and we are running eg "tmux refresh-client" already in
things like prompt_command() so let's avoid another fork+exec

for the server, it might still be a good idea to use the path, still
debating, for now we'll just keep it as 'tmux' and see if we like it.
the real path is always available via the procfile if user needs to see
how it got started, but he should know already based on his PATH

exec -a was added early, checking commit messages in bash repo, it was
before 2.0, stopped looking there because that was early days, there
will never be an extant system we need to run on that's that old and
there would be a whole lot of other stuff that broke too if we had to