- rc: update comments with new info re huge pages and posix_spawn()
we have a bunch of issues related to '#()', "refresh-client -S", new
windows and scrollback, which we attempt to mitigate with huge pages,
and have a future fix possible with posix_spawn (actually we are
starting to run with a patch which partially fixes this even now).
update the comments to reflect new information.