^ tmux: malloc: trim_threshold remove this tunable to reduce cpu
our issue is not with memory but with contiguous memory and cpu usage to
find it.  no need to add to it by reducing the trim threshold.
typically we have enough memory it's just fragmented so THP takes a
while to find it and/or gives up and defers until kcompactd does
something and it's always behind the demand when we have sudden large
allocations in tmux