^ v2rc: bind: paste: disable bracketed pastes, more trouble than worth
this bracketed paste thing is routinely causing issues (like, it
sporadically does not work and we see the escapes), am disabling it in
inputrc, vim and tmux.  this could very well be an issue with my
terminal (st), but looking at its code, it does support the sequences.
unclear why it doesn't work.  anyways we don't need it and lived just
fine without it for decades so we're just going to disable it and
revisit later if we ever want to.